requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
# Core Streamlit and web framework
streamlit>=1.28.0

# Google APIs for Gmail
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
google-auth>=2.23.0

# HTTP requests for Ollama API
requests>=2.31.0

# Machine Learning and NLP
scikit-learn>=1.3.0
numpy>=1.24.0

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Data handling
pandas>=2.0.0

# Optional: Enhanced text processing
nltk>=3.8.1
sentence-transformers>=2.2.2

# Development and debugging
python-dotenv>=1.0.0